requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
asyncpg>=0.29.0
pyjwt>=2.8.0
argon2-cffi>=23.1.0
pydantic[email]>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aioredis>=2.0.1
celery>=5.3.6
setuptools>=65.0.0
PyYAML>=6.0
akshare>=1.12.0
tushare>=1.2.89
matplotlib>=3.7.0
openpyxl>=3.1.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
